+++ packages/next/src/provider/ClientProviderWrapper.tsx
@@ -2,9 +2,9 @@
 import { ClientProvider as _ClientProvider } from 'gt-react/client';
 import { ClientProviderProps } from 'gt-react/internal';
 import { usePathname } from 'next/navigation';
-import { useEffect } from 'react';
+import { useEffect, useMemo } from 'react';
 import { extractLocale } from '../middleware-dir/utils';
-import { determineLocale, standardizeLocale } from 'generaltranslation';
+import { GT, standardizeLocale } from 'generaltranslation';
 
 export default function ClientProvider(
   props: ClientProviderProps & {
@@ -19,7 +19,24 @@ export default function ClientProvider(
     gtServicesEnabled,
     referrerLocaleCookieName,
     localeRoutingEnabledCookieName,
+    devApiKey,
+    projectId,
+    runtimeUrl,
+    customMapping,
   } = props;
+
+  const gt = useMemo(
+    () =>
+      new GT({
+        devApiKey,
+        sourceLocale: defaultLocale,
+        targetLocale: locale,
+        projectId,
+        baseUrl: runtimeUrl || undefined,
+        customMapping,
+      }),
+    [devApiKey, defaultLocale, locale, projectId, runtimeUrl, customMapping]
+  );
 
   // Trigger page reload when locale changes
   // When nav to same route but in diff locale, client components were cached and not re-rendered
@@ -27,7 +44,7 @@ export default function ClientProvider(
   useEffect(() => {
     // ----- Referrer Locale ----- //
     if (locale) {
-      document.cookie = `${referrerLocaleCookieName}=${locale};path=/`;
+      document.cookie = `${referrerLocaleCookieName}=${gt.resolveAliasLocale(locale)};path=/`;
     }
 
     // ----- Middleware ----- //
@@ -40,8 +57,8 @@ export default function ClientProvider(
         ?.split('=')[1] === 'true';
     if (middlewareEnabled) {
       // Extract locale from pathname
-      const extractedLocale = extractLocale(pathname) || defaultLocale;
-      const pathLocale = determineLocale(
+      const extractedLocale = extractLocale(pathname, gt) || defaultLocale;
+      let pathLocale = gt.determineLocale(
         [
           gtServicesEnabled
             ? standardizeLocale(extractedLocale)
@@ -50,6 +67,9 @@ export default function ClientProvider(
         ],
         locales
       );
+      if (pathLocale) {
+        pathLocale = gt.resolveAliasLocale(pathLocale);
+      }
 
       if (pathLocale && locales.includes(pathLocale) && pathLocale !== locale) {
         // clear cookie (avoids infinite loop when there is no middleware)