+++ packages/react/src/provider/hooks/locales/useLocaleState.ts
@@ -1,6 +1,15 @@
 import { useMemo } from 'react';
 import { useDetermineLocale } from './useDetermineLocale';
-import { requiresTranslation, isSameLanguage } from 'generaltranslation';
+import {
+  requiresTranslation,
+  isSameLanguage,
+  isValidLocale,
+} from 'generaltranslation';
+import { CustomMapping } from 'generaltranslation/types';
+import {
+  invalidCanonicalLocalesError,
+  invalidLocalesError,
+} from '../../../errors/createErrors';
 
 export function useLocaleState({
   _locale,
@@ -8,12 +17,14 @@ export function useLocaleState({
   locales: _locales,
   ssr,
   localeCookieName,
+  customMapping,
 }: {
   _locale: string;
   defaultLocale: string;
   locales: string[];
   ssr: boolean;
   localeCookieName: string;
+  customMapping?: CustomMapping;
 }) {
   // Locale standardization
   const locales = useMemo(() => {
@@ -26,6 +37,7 @@ export function useLocaleState({
     locales,
     ssr,
     localeCookieName,
+    customMapping,
   });
 
   const [translationRequired, dialectTranslationRequired] = useMemo(() => {
@@ -33,12 +45,39 @@ export function useLocaleState({
     const translationRequired = requiresTranslation(
       defaultLocale,
       locale,
-      locales
+      locales,
+      customMapping
     );
 
     // User locale is not default locale but is a dialect of the same language
     const dialectTranslationRequired =
       translationRequired && isSameLanguage(defaultLocale, locale);
+
+    // Check: invalid locale
+    if (!customMapping) {
+      const invalidLocales: string[] = [];
+      locales.forEach((locale) => {
+        if (!isValidLocale(locale)) {
+          invalidLocales.push(locale);
+        }
+      });
+      if (invalidLocales.length) {
+        throw new Error(invalidLocalesError(invalidLocales));
+      }
+    }
+
+    // Check: invalid canonical locale
+    if (customMapping) {
+      const invalidCanonicalLocales: string[] = [];
+      locales.forEach((locale) => {
+        if (!isValidLocale(locale, customMapping)) {
+          invalidCanonicalLocales.push(locale);
+        }
+      });
+      if (invalidCanonicalLocales.length) {
+        throw new Error(invalidCanonicalLocalesError(invalidCanonicalLocales));
+      }
+    }
 
     return [translationRequired, dialectTranslationRequired];
   }, [defaultLocale, locale, locales]);